[Codegen] Implement PipelineAttrInterface directly for GPU/SPIRV
Move callback storage and buildPipeline/emitConstraints into IREEGPUAttrs
directly. Suggested by Hanhan in
https://github.com/iree-org/iree/pull/23864#issuecomment-4095024073.
