+++ compiler/src/iree/compiler/Codegen/Dialect/GPU/ExternalInterfaces/BUILD.bazel
@@ -15,16 +15,12 @@ package(
 iree_compiler_cc_library(
     name = "ExternalModels",
     srcs = [
-        "GPUPipelineExternalModels.cpp",
         "GPUScopeExternalModels.cpp",
         "Interfaces.cpp",
-        "SPIRVPipelineExternalModels.cpp",
     ],
     hdrs = [
-        "GPUPipelineExternalModels.h",
         "GPUScopeExternalModels.h",
         "Interfaces.h",
-        "SPIRVPipelineExternalModels.h",
     ],
     deps = [
         "//compiler/src/iree/compiler/Codegen/Dialect/Codegen/IR:IREECodegenDialect",
