+++ compiler/src/iree/compiler/Codegen/Dialect/GPU/ExternalInterfaces/CMakeLists.txt
@@ -14,15 +14,11 @@ iree_cc_library(
   NAME
     ExternalModels
   HDRS
-    "GPUPipelineExternalModels.h"
     "GPUScopeExternalModels.h"
     "Interfaces.h"
-    "SPIRVPipelineExternalModels.h"
   SRCS
-    "GPUPipelineExternalModels.cpp"
     "GPUScopeExternalModels.cpp"
     "Interfaces.cpp"
-    "SPIRVPipelineExternalModels.cpp"
   DEPS
     LLVMSupport
     MLIRArithDialect
